add linter
