--- src/index.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
-
-// https://www.figma.com/file/5Up611L8QTVThfzfi9fwI8/%D0%9A%D1%83%D0%B4%D0%B0-%D0%9F%D0%B8%D1%86%D1%86%D0%B0-(Copy)-(Copy)?type=design&t=jOMbmStDRmNqJjxE-6
-reportWebVitals();
